job/requirements.txt
astroid==2.1.0
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2
djangorestframework
elasticsearch==6.3.1
gunicorn==19.8.1
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mailjet-rest==1.3.0
mccabe==0.6.1
psycopg2-binary
pylint==2.2.2
pytz==2018.4
requests==2.18.4
six==1.12.0
sqlparse==0.3.0
typed-ast==1.1.1
urllib3==1.22
whitenoise==4.1.2
wrapt==1.10.11
